+++ install/5_install_core_apps.sh
@@ -62,5 +62,5 @@ else
 	. SCRIPTS/inst_core_apps.sh
 fi
 
-info "Core Apps have been installed. You should now execute stage 6: Settings"
+info "Core Apps have been installed. You should now execute next stage: Settings"